Moved the ResourceLoaderDelegate to its own separate file
